Stop checking if volume to mount in k8s exists locally
It's entirely possible (and not an error) to create a podspec that
contains paths to volume-mount that do not exist wherever
service_configuration_lib is being invoked from.

As long as we ensure that we don't double-add things (which is only a
problem due to the way that the env var keys are constructed here),
skipping the existence check should be fine.