Remove default mounting of /etc/pki/spark
I missed that this was a thing that we did in my previous PR to always
set k8s mounts regardless of whether or not they exist on the host
executing service_configuration_lib code - however, my assertion that
in the worst case the container runtime would create the missing files
wherever these mounts are used was incorrect: in a `paasta spark-run`,
the spark driver will run locally and re-use the k8s volume functions to
figure out what needs to be mounted.

This would normally be fine, but we have a security setup that prevents
writes at certain paths: of which /etc/pki is in the set of blocked
paths.

Since we no longer have a spark cluster that is able to use
certificate-based k8s authentication, this should be totally safe to
remove as a default.